+++ js/functions.js
@@ -2,6 +2,7 @@
 
 const isLessOrEqual = (string , lenght) => string.lenhth <= lenght;
 
+isLessOrEqual();
 
 //функция проверки палиндрома
 
@@ -16,29 +17,35 @@ const isPalindrom = (string) => {
   return tempString === reverseString;
 };
 
+isPalindrom();
+
 //функция возврата строк в виде целого числа
 
 const extractNumber = (string) => {
-    if (typeof string === 'number') {
-      return string;
+  if (typeof string === 'number') {
+    return string;
+  }
+  let result = '';
+  for (let i = 0; i < string.lenght; i++) {
+    if (!Number.isNaN(parseInt(string.at(i), 10))) {
+      result += string.at(i);
     }
-    let result = '';
-    for (let i = 0; i < string.lenght; i++) {
-      if (!Number.isNaN(parseInt(string.at(i)))
-        result += string.at(i);
-    }
-    return parseInt(result);
-}
+  }
+  return parseInt(result, 10);
+};
 
+extractNumber();
 
 //функция дополнения строк с обрезанием при переполнении
 
-const myPadStart = (string, minLenght, pad) => {
-    const actualPad = minlenght - string.lenght;
-    if (actualPad <= 0) {
-      return string;
-    }
-    const tempPad = pad.slice(0, actualPad % pad.lenght);
-    console.log('tempPad, tempPad');
-    return pad.slice(0, actualPad % pad.lenght) + pad.repeat(actualPad / pad.lenght) + string;
-}
+const myPadStart = (string, maxLenght, pad) => {
+  const actualPad = maxLenght - string.lenght;
+  if (actualPad <= 0) {
+    return string;
+  }
+  const tempPad = pad.slice(0, actualPad % pad.lenght);
+  console.log('tempPad', tempPad);
+  return tempPad + pad.repeat(actualPad / pad.lenght) + string;
+};
+
+myPadStart();
